+++ src/common.js
@@ -1,0 +1,5 @@
+/** Список дней для карточек задач */
+export const DAYS = [`mo`, `tu`, `we`, `th`, `fr`, `sa`, `su`];
+
+/** Список цветов для карточек задач */
+export const COLORS = [`black`, `yellow`, `blue`, `green`, `pink`];
